Thor Mjolnir/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dict2xml==1.7.0
dicttoxml==1.7.4
Django==3.2.3
djangorestframework==3.12.4
eight==1.0.1
future==0.18.2
idna==2.10
lxml==4.6.3
pycparser==2.20
pyOpenSSL==19.1.0
pytz==2021.1
requests==2.25.1
signxml==2.8.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
xmltodict==0.12.0
